refactor: Replace DevServerPlugin with a top-level function
