--- src/main/kotlin/server/DevServerPlugin.kt
@@ -1,107 +1,0 @@
-package ru.endlesscode.bukkitgradle.server
-
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.gradle.api.file.Directory
-import org.gradle.api.provider.Provider
-import org.gradle.api.tasks.TaskContainer
-import org.gradle.api.tasks.TaskProvider
-import org.gradle.jvm.toolchain.JavaToolchainService
-import org.gradle.kotlin.dsl.*
-import ru.endlesscode.bukkitgradle.Bukkit
-import ru.endlesscode.bukkitgradle.bukkit
-import ru.endlesscode.bukkitgradle.extensions.java
-import ru.endlesscode.bukkitgradle.plugin.util.resolveMinimalJavaVersion
-import ru.endlesscode.bukkitgradle.server.extension.ServerConfiguration
-import ru.endlesscode.bukkitgradle.server.task.CreateIdeaGradleRunConfiguration
-import ru.endlesscode.bukkitgradle.server.task.PrepareServer
-import xyz.jpenilla.runpaper.RunPaperPlugin
-import xyz.jpenilla.runpaper.task.RunServer
-import java.io.File
-
-public class DevServerPlugin : Plugin<Project> {
-
-    private lateinit var project: Project
-    private lateinit var bukkit: Bukkit
-
-    private val serverConfiguration: ServerConfiguration
-        get() = bukkit.server
-
-    private val tasks: TaskContainer
-        get() = project.tasks
-
-    @Override
-    override fun apply(target: Project) {
-        project = target
-        bukkit = project.bukkit
-
-        target.apply<RunPaperPlugin>()
-        val configuredServerDir = target.resolveConfiguredServerDir()
-
-        // Preconfigure RunServer task
-        val serverVersion = project.provider<String> { serverConfiguration.version }.orElse(bukkit.apiVersion)
-        val runServer = tasks.named<RunServer>("runServer") {
-            version.convention(serverVersion)
-            if (configuredServerDir != null) runDirectory.convention(configuredServerDir)
-            jvmArgs(serverConfiguration.buildJvmArgs())
-            args(serverConfiguration.bukkitArgs)
-
-            defaultCharacterEncoding = serverConfiguration.encoding
-            debugOptions {
-                enabled.convention(serverConfiguration.debug)
-                suspend.convention(false)
-            }
-        }
-
-        // RunPaperPlugin uses afterEvaluate under the hood, so we have to use afterEvaluate
-        // to set our conventions after their ones
-        project.afterEvaluate { configureDefaultJvmForServer() }
-
-        val prepareServer = registerPrepareServerTask(runServer)
-        runServer.configure { dependsOn(prepareServer) }
-
-        registerBuildIdeRunTask(runServer)
-    }
-
-    private fun Project.configureDefaultJvmForServer() {
-        val toolchains = project.extensions.findByType<JavaToolchainService>() ?: return
-        val spec = java.toolchain
-
-        tasks.withType<RunServer>().configureEach {
-            javaLauncher.convention(
-                toolchains.launcherFor {
-                    languageVersion.convention(version.map(::resolveMinimalJavaVersion))
-                    implementation.convention(spec.implementation)
-                    vendor.convention(spec.vendor)
-                }
-            )
-        }
-    }
-
-    private fun Project.resolveConfiguredServerDir(): Provider<Directory>? {
-        val deprecated = DeprecatedServerProperties(rootDir, providers)
-
-        val serverDirProperty = providers.gradleProperty("bukkitgradle.server.dir")
-            .orElse(provider { deprecated.devServerDir?.absolutePath })
-            .orNull ?: return null
-        val serverDirFile = provider { File(serverDirProperty).absoluteFile }
-        return layout.dir(serverDirFile)
-    }
-
-    private fun registerPrepareServerTask(
-        runServer: Provider<RunServer>,
-    ): TaskProvider<PrepareServer> {
-        return tasks.register<PrepareServer>("prepareServer") {
-            this.serverDir.set(runServer.map { it.runDirectory.get() })
-            eula = serverConfiguration.eula
-            onlineMode = serverConfiguration.onlineMode
-        }
-    }
-
-    private fun registerBuildIdeRunTask(runServer: TaskProvider<RunServer>) {
-        tasks.register<CreateIdeaGradleRunConfiguration>("buildIdeaRun") {
-            configurationName.set("Run Server [${project.name}]")
-            taskNames.add(runServer.name)
-        }
-    }
-}
